Restaurant/serializers.py
from rest_framework import serializers
from .models import Restaurant, Menu

class RestaurantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Restaurant
        fields = ['restaurant_id', 'name', 'address', 'taste']

class MenuSerializer(serializers.ModelSerializer):
    restaurant = RestaurantSerializer(read_only=True)
    
    class Meta:
        model = Menu
        fields = ['menu_id', 'name', 'price', 'imageurl', 'restaurant']
